requirements.txt
boto3>=1.28.0
bcrypt>=4.0.0
qrcode[pil]>=7.4.0
Pillow>=10.0.0
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
bcrypt
flask-cors
pillow
qrcode
flask-login
python-dotenv
qrcode
pyjwt
awsgi
pyjwt
awsgi
google-cloud-firestore
google-cloud-storage
google-generativeai
pytest
flask-sock
librosa
soundfile
google-genai
sendgrid
mangum
sift-stack-py
